Move libocto ir prompts into libocto, and move result.ts into libocto
