+++ source/compilers/optimize-files.ts
@@ -1,7 +1,7 @@
 import type { octoAgent } from "../ir/octo-ir.ts";
 import type { LoweredIRWithTrajectories } from "../libocto/llm-ir.ts";
 import type { ToolCall } from "../libocto/tool-def.ts";
-import * as irPrompts from "../prompts/ir-prompts.ts";
+import * as irPrompts from "../prompts/octo-ir-prompts.ts";
 import { canDisplayImage } from "../providers.ts";
 import type { MultimodalConfig } from "../providers.ts";
 import type { FileMutateIR, FileReadIR } from "../tools/common.ts";
