+++ source/libocto/result.ts
