+++ source/prompts/octo-ir-prompts.ts
@@ -1,16 +1,10 @@
 import { ImageInfo } from "../utils/image-utils.ts";
 import { CanDisplayImageResult } from "../providers.ts";
+import { imageAttachmentPlaceholderText } from "../libocto/compilers/ir-prompts.ts";
 
 export function toolReject() {
   return `
 Tool call was rejected by user. Your tool call did not run. No changes were applied.
-`.trim();
-}
-
-export function toolSkip(reason: string) {
-  return `
-Tool was skipped and didn't run. The reason for skipping the tool was:
-${reason}
 `.trim();
 }
 
@@ -34,8 +28,4 @@ export function imageAttachmentPlaceholder(content: string, images: ImageInfo[])
   if (images.length === 0) return content;
   const imageBlurbs = images.map(() => imageAttachmentPlaceholderText()).join("\n");
   return `${imageBlurbs}\n\n${content}`;
-}
-
-export function imageAttachmentPlaceholderText() {
-  return "[An image was attached here. Since images are not supported by your model, the source to the image is omitted. There might be future context that allows you to make a guess about what the image was, so keep that in mind as you process the rest of the messages.]";
 }
