+++ source/tools/lsp-common.ts
@@ -4,7 +4,7 @@ import { LspClient } from "../lsp/client.ts";
 import type { Config } from "../config.ts";
 import type { Transport } from "../transports/transport-common.ts";
 import { getLspClientForFile } from "../lsp/detect.ts";
-import { Result, ok, err, attempt } from "../result.ts";
+import { Result, ok, err, attempt } from "../libocto/result.ts";
 
 export type LspFileOnlyArgs = t.GetType<typeof LspFileOnlyArgumentsSchema>;
 export const LspFileOnlyArgumentsSchema = t.subtype({
