+++ source/tools/tool-defs/grep.ts
@@ -4,7 +4,7 @@ import { TOOL, USER_ABORTED_ERROR_MESSAGE } from "../common.ts";
 import { getModelFromConfig } from "../../config.ts";
 import { AbortError, CommandFailedError } from "../../transports/transport-common.ts";
 import { estimateTokens } from "../../ir/count-ir-tokens.ts";
-import { ok, err, toErrString } from "../../result.ts";
+import { ok, err, toErrString } from "../../libocto/result.ts";
 
 export default TOOL.declare({
   name: "grep",
