+++ source/tools/tool-defs/list.ts
@@ -1,7 +1,7 @@
 import { t } from "structural";
 import { TOOL } from "../common.ts";
 import { Transport } from "../../transports/transport-common.ts";
-import { ok, err, attempt } from "../../result.ts";
+import { ok, err, attempt } from "../../libocto/result.ts";
 
 const ArgumentsSchema = t.subtype({
   dirPath: t.optional(t.str.comment("Path to the directory")),
