+++ source/tools/tool-defs/prepend.ts
@@ -7,7 +7,7 @@ import {
   fileMutateIR,
 } from "../common.ts";
 import { Transport } from "../../transports/transport-common.ts";
-import { ok, err } from "../../result.ts";
+import { ok, err } from "../../libocto/result.ts";
 
 const ArgumentsSchema = t.subtype({
   filePath: t.str.comment("The path to the file"),
